refactor: extract dodge reset stats accumulator
